+++ samplesheets/src/components/Overview.vue
@@ -63,6 +63,9 @@
         <list-row :legend="'Title'"
                   :value="studyInfo['display_name']">
         </list-row>
+        <list-row :legend="'Identifier'"
+                  :value="studyInfo['identifier']">
+        </list-row>
         <list-row :legend="'Description'"
                   :value="studyInfo['description']">
         </list-row>